+++ src/App.tsx
@@ -2,11 +2,13 @@ import { useState, useEffect, useRef, useCallback, useMemo, type CSSProperties }
 import { AnimatePresence, motion } from 'framer-motion'
 import { resolveStage } from './stages'
 import { SettingsProvider } from './hooks/useSettings'
-import { PortalProvider, usePortal } from './context/PortalContext'
-import { StreamingProvider, useStreaming } from './context/StreamingContext'
+import { PortalProvider } from './context/PortalContext'
+import { usePortal } from './context/portalContextValue'
+import { StreamingProvider } from './context/StreamingContext'
+import { useStreaming } from './context/streamingContextValue'
 import { VortexProvider } from './context/VortexContext'
 import { AudioProvider } from './context/AudioContext'
-import { useAudio } from './context/AudioContext'
+import { useAudio } from './context/audioContextValue'
 import AudioController from './components/AudioController'
 import { useAppStartup } from './hooks/useAppStartup'
 import { invoke } from './bridge'
