+++ src/components/ConnectionLostOverlay.tsx
@@ -1,7 +1,7 @@
 import { useCallback } from 'react'
 import { buildDiagnosticsPayload } from '../lib/diagnosticsPayload'
-import { useStreaming } from '../context/StreamingContext'
-import { useSettings } from '../hooks/useSettings'
+import { useStreaming } from '../context/streamingContextValue'
+import { useSettings } from '../hooks/settingsContextValue'
 import Button from './ui/Button'
 import ServerLogDisplay from './ServerLogDisplay'
 import { useTranslation } from 'react-i18next'
