+++ src/components/MenuSettingsView.tsx
@@ -2,8 +2,8 @@ import { useState, useEffect, useCallback, useRef } from 'react'
 import { useTranslation } from 'react-i18next'
 import type { TranslationKey } from '../i18n'
 import { VIEW_DESCRIPTION, VIEW_HEADING } from '../styles'
-import { useSettings } from '../hooks/useSettings'
-import { useStreaming } from '../context/StreamingContext'
+import { useSettings } from '../hooks/settingsContextValue'
+import { useStreaming } from '../context/streamingContextValue'
 import { useVolumeControls } from '../hooks/useVolumeControls'
 import MenuButton from './ui/MenuButton'
 import SettingsToggle from './ui/SettingsToggle'
