+++ src/components/TerminalDisplay.tsx
@@ -2,9 +2,9 @@ import { useCallback, useEffect, useMemo, useState } from 'react'
 import { invoke } from '../bridge'
 import { buildDiagnosticsPayload } from '../lib/diagnosticsPayload'
 import { resolveStage } from '../stages'
-import { useStreaming } from '../context/StreamingContext'
-import { useVortex } from '../context/VortexContext'
-import { useSettings } from '../hooks/useSettings'
+import { useStreaming } from '../context/streamingContextValue'
+import { useVortex } from '../context/vortexContextValue'
+import { useSettings } from '../hooks/settingsContextValue'
 import { useEngineLogs } from '../hooks/useEngineLogs'
 import Button from './ui/Button'
 import { GooseFactTicker } from './GooseMode'
