+++ src/components/settings/DebugTab.tsx
@@ -3,7 +3,7 @@ import { useTranslation } from 'react-i18next'
 import { SETTINGS_MUTED_TEXT } from '../../styles'
 import { buildDiagnosticsPayload } from '../../lib/diagnosticsPayload'
 import { ENGINE_MODES, type Settings } from '../../types/settings'
-import { useStreaming } from '../../context/StreamingContext'
+import { useStreaming } from '../../context/streamingContextValue'
 import SettingsSection from '../ui/SettingsSection'
 import SettingsRow from '../ui/SettingsRow'
 import SettingsCheckbox from '../ui/SettingsCheckbox'
