+++ src/components/settings/EngineTab.tsx
@@ -3,8 +3,8 @@ import { useTranslation } from 'react-i18next'
 import { invoke } from '../../bridge'
 import { SETTINGS_MUTED_TEXT } from '../../styles'
 import { ENGINE_MODES, QUANT_OPTIONS, type QuantOption, type Settings } from '../../types/settings'
-import { useSettings } from '../../hooks/useSettings'
-import { useStreaming } from '../../context/StreamingContext'
+import { useSettings } from '../../hooks/settingsContextValue'
+import { useStreaming } from '../../context/streamingContextValue'
 import { normalizeServerUrl, toHealthUrl } from '../../utils/serverUrl'
 import SettingsSection from '../ui/SettingsSection'
 import SettingsToggle from '../ui/SettingsToggle'
