+++ src/context/FocusScopeContext.tsx
@@ -1,28 +1,6 @@
-/** Focus-scope stack for gamepad/keyboard navigation. Modals push a scope on
- *  mount so spatial navigation and the B=back handler are contained to them,
- *  and pop on unmount (restoring the previously-focused element). */
-
 import { useEffect, useRef, type ReactNode } from 'react'
 import { findFirstFocusable } from '../lib/focusNavigation'
-
-export type FocusScope = {
-  root: HTMLElement
-  getOnCancel: () => (() => void) | undefined
-  previousFocus: Element | null
-}
-
-const stack: FocusScope[] = []
-
-export const getTopFocusScope = (): FocusScope | null => stack[stack.length - 1] ?? null
-
-/** Scope used when the stack is empty — spatial navigation searches the whole document. */
-export const getActiveScopeRoot = (): ParentNode => getTopFocusScope()?.root ?? document
-
-const pushScope = (s: FocusScope) => stack.push(s)
-const popScope = (s: FocusScope) => {
-  const idx = stack.lastIndexOf(s)
-  if (idx >= 0) stack.splice(idx, 1)
-}
+import { pushScope, popScope, type FocusScope } from './focusScopeStack'
 
 type FocusScopeProps = {
   children: ReactNode
